Project Setup, Authentication & Backend Setup (#5)
Sets up the core tools that will be used in this project as well as a
template of the project structure that can be mimicked for future
expansion. Firebase has been integrated as the backend service along
with the relevant encryption mechanisms to keep the data secure
online. The Authentication Page has been done up as well, some minor
tweaks are still possible but this will suffice for a first iteration.